moving game setup to game_loop, fixing small bug in simulator that was overriding num_players
